Move LUA tests to a separate file
