+++ test/test_aioredis1.py
@@ -1,18 +1,25 @@
 import asyncio
 
+import pytest
 import pytest_asyncio
 from packaging.version import Version
-import pytest
+
 aioredis = pytest.importorskip("aioredis")
 
 import fakeredis.aioredis
-
 
 aioredis2 = Version(aioredis.__version__) >= Version('2.0.0a1')
 pytestmark = [
     pytest.mark.asyncio,
     pytest.mark.skipif(aioredis2, reason="Test is only applicable to aioredis 1.x")
 ]
+import importlib
+
+lua_module = importlib.util.find_spec("lupa")
+lupa_required_for_test = pytest.mark.skipif(
+    lua_module is None,
+    reason="Test is only applicable if lupa is installed"
+)
 
 
 @pytest_asyncio.fixture(
@@ -74,7 +81,7 @@ async def test_transaction_fail(r, conn):
     # ensure that the WATCH applies to the same connection as the MULTI/EXEC.
     await r.set('foo', '1')
     await conn.watch('foo')
-    await conn.set('foo', '2')    # Different connection
+    await conn.set('foo', '2')  # Different connection
     tr = conn.multi_exec()
     tr.get('foo')
     with pytest.raises(aioredis.MultiExecError):
@@ -117,6 +124,7 @@ async def test_blocking_timeout(conn):
 @pytest.mark.slow
 async def test_blocking_unblock(r, conn, event_loop):
     """Blocking command that gets unblocked after some time."""
+
     async def unblock():
         await asyncio.sleep(0.1)
         await r.rpush('list', 'y')
@@ -153,6 +161,7 @@ async def test_no_script_error(r):
         await r.evalsha('0123456789abcdef0123456789abcdef')
 
 
+@lupa_required_for_test
 async def test_failed_script_error(r):
     await r.set('foo', 'bar')
     with pytest.raises(aioredis.ReplyError, match='^ERR Error running script'):
